refactor: consolidate trade validator
